Refactor: 로그인부분(계층별 코드 정리)
